JogoDaVelha

+++ src/Main.java
@@ -1,9 +1,15 @@
 package src;
+
+// tudo desenvolvido no programa, vai dentro dessa class
 public class Main {
     public static void main(String[] args) {
-        int val1 = 10;
-        int val2 = 25;
-        double total = val1 +val2;
+        // byte / short / int / long / sao formas de declara uma variavel numerica
+        // float / double sao numeros com ponto flutuante
+        // boolean / valor verdadeiro
+        // char / valor caracter
+        byte val1 = 10;
+        short val2 = 25;
+        int total = val1 +val2;
         System.out.println(total);
 
     }
